refactor: Answer의 일급 컬렉션 생성
